+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "tcg" extension will be documented in this file.
 
+## [1.0.0]
+
+- Logic was changed to improve the perfomance of the tracking-code replacement,
+- Now extension will ask you to execute two main action (analyze, rewrite)
+  -analyze will look into the code to an existing code and will start the count from the biggest number
+  -rewrite will overwrite the tracking-codes in the whole file, rewrite has the same old logic to generate tracking-code
+
 ## [0.0.3]
 
 - Added error handling when aborting the creation of the tracking-code